% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandextexttranslater.R
\name{yandex_detect_language}
\alias{yandex_detect_language}
\title{Detects the language of the specified text.}
\usage{
yandex_detect_language(yandex_api_key, text = "")
}
\arguments{
\item{yandex_api_key}{yandex API key}

\item{text}{The text to detect the language for}
}
\value{
String of detected language (character type)
See \url{https://tech.yandex.com/translate/doc/dg/reference/detect-docpage/} for more details
}
\description{
Detects the language of the specified text.
}
\examples{
detected_languaged = yandex_detect_language(yandex_api_key, text="voglio mangiare cena")
}
