% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandextexttranslater.R
\name{yandex_supported_languages}
\alias{yandex_supported_languages}
\title{Gets a list of translation directions supported by the service}
\usage{
yandex_supported_languages(yandex_api_key, lang = "en")
}
\arguments{
\item{yandex_api_key}{yandex API key}

\item{lang}{Defaults to English ("en")}
}
\value{
A list with two data frames
The first data frame gives the supported translation directions (all columns are character data types)
The second data frame matches language abbreviations with full language name (all columns are character data types)
See \url{https://tech.yandex.com/translate/doc/dg/reference/getLangs-docpage/} for more details
}
\description{
Gets a list of translation directions supported by the service
}
\examples{
supported_languages = yandex_supported_languages(yandex_api_key)
available_translations = supported_languages[[1]]
abbreviations = supported_languages[[2]]
}
