% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandextexttranslater.R
\name{yandex_translate}
\alias{yandex_translate}
\title{Translates text to the specified language}
\usage{
yandex_translate(yandex_api_key, text = "", lang = "")
}
\arguments{
\item{yandex_api_key}{yandex API key}

\item{text}{The text to translate. The maximum size of the text being passed is 10000 characters.}

\item{lang}{The translation direction. You can use any of the following ways to set it:
As a pair of language codes separated by a hyphen ("from"-"to").
For example, en-ru indicates translating from English to Russian.
As the final language code (for example, ru). In this case, the service tries to detect the source language automatically.}
}
\value{
translated text (character type)
See \url{https://tech.yandex.com/translate/doc/dg/reference/translate-docpage/} for more details
}
\description{
Translates text to the specified language
}
\examples{
translated_text = yandex_translate(yandex_api_key, text="voglio mangiare cena", lang="it-en")
}
